+++ llmfoundry/models/inference_api_wrapper/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2022 MosaicML LLM Foundry authors
 # SPDX-License-Identifier: Apache-2.0
 
-from llmfoundry.models.inference_api_wrapper.fmapi_lm import (
+from llmfoundry.models.inference_api_wrapper.fmapi import (
     FMAPICasualLMEvalWrapper, FMAPIChatAPIEvalWrapper)
 from llmfoundry.models.inference_api_wrapper.interface import \
     InferenceAPIEvalWrapper
